Version 0.5
In this version the application have not only the interface developed but too the conection for a local database and too an insertion SQL command. Probably in the next days this application will be ended with the addition of SELECT, UPDATE and DELETE. Thank you soo much for my teacher of System Development Ms. Marlene da Silva Maximiano de Oliveira for this content teached for us.